Bump to 0.6.0 for release
This release bumps ESLint to the 8.x series, prettier to the 3.x series,
and TypeScript to the 5.x series. It also bumps other dependencies to
their latest versions and is now self-applying this lint config in CI.

Note that one big and impactful change is the switch to Prettier 3.x,
which [changes the default value of `trailingComma` to
`all`](https://prettier.io/blog/2023/07/05/3.0.0.html#javascript). This
means that any project that pulls in a release after this PR lands will
likely have a lot of updates to function calls that were previously not
using trailing commas.

Somewhat less invasive, TypeScript annotations like `typeof
Thing["key"]` are now being formatted as `(typeof Thing)["key"]`.